refactor(disagg): merge disagg-headers-handler PreRequest into disagg-profile-handler
Handler now implements both ProfileHandler and PreRequest, so configs
no longer need a separate disagg-headers-handler entry.

- Move PreRequest method and constants into disagg_profile_handler.go
- Delete disagg_headers_handler.go and its test file
- Register legacy type strings (disagg-headers-handler,
  prefill-header-handler) as no-op factories that log a deprecation
  warning, preserving backward compatibility for old YAML configs
- Remove disagg-headers-handler entries from deploy configs, e2e test
  configs, and documentation

Closes #833

Made-with: Cursor
