+++ docs/disaggregation.md
@@ -277,7 +277,6 @@ plugins:
       maxPrefixBlocksToMatch: 256
       lruCapacityPerServer: 31250
   - type: max-score-picker
-  - type: disagg-headers-handler
   - type: prefix-based-pd-decider
     parameters:
       nonCachedTokens: 8
@@ -336,7 +335,6 @@ plugins:
       maxPrefixBlocksToMatch: 256
       lruCapacityPerServer: 31250
   - type: max-score-picker
-  - type: disagg-headers-handler
   - type: always-disagg-multimodal-decider
   - type: prefix-based-pd-decider
     parameters:
